refactor: table instructions
